+++ clients/common/lapack/testing_getf2_getrf.hpp
@@ -35,21 +35,21 @@
 #include "common/misc/rocsolver_arguments.hpp"
 #include "common/misc/rocsolver_test.hpp"
 
-template <bool STRIDED, bool GETRF, typename T, typename U>
+template <bool STRIDED, bool GETRF, typename I, typename Td, typename Id>
 void getf2_getrf_checkBadArgs(const rocblas_handle handle,
-                              const rocblas_int m,
-                              const rocblas_int n,
-                              T dA,
-                              const rocblas_int lda,
+                              const I m,
+                              const I n,
+                              Td dA,
+                              const I lda,
                               const rocblas_stride stA,
-                              U dIpiv,
+                              Id dIpiv,
                               const rocblas_stride stP,
-                              U dinfo,
-                              const rocblas_int bc)
+                              Id dInfo,
+                              const I bc)
 {
     // handle
     EXPECT_ROCBLAS_STATUS(
-        rocsolver_getf2_getrf(STRIDED, GETRF, nullptr, m, n, dA, lda, stA, dIpiv, stP, dinfo, bc),
+        rocsolver_getf2_getrf(STRIDED, GETRF, nullptr, m, n, dA, lda, stA, dIpiv, stP, dInfo, bc),
         rocblas_status_invalid_handle);
 
     // values
@@ -58,57 +58,57 @@ void getf2_getrf_checkBadArgs(const rocblas_handle handle,
     // sizes (only check batch_count if applicable)
     if(STRIDED)
         EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, dA, lda, stA,
-                                                    dIpiv, stP, dinfo, -1),
+                                                    dIpiv, stP, dInfo, -1),
                               rocblas_status_invalid_size);
 
     // pointers
-    EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, (T) nullptr, lda, stA,
-                                                dIpiv, stP, dinfo, bc),
+    EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, (Td) nullptr, lda,
+                                                stA, dIpiv, stP, dInfo, bc),
                           rocblas_status_invalid_pointer);
     EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, dA, lda, stA,
-                                                (U) nullptr, stP, dinfo, bc),
+                                                (Id) nullptr, stP, dInfo, bc),
                           rocblas_status_invalid_pointer);
     EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, dA, lda, stA, dIpiv,
-                                                stP, (U) nullptr, bc),
+                                                stP, (Id) nullptr, bc),
                           rocblas_status_invalid_pointer);
 
     // quick return with invalid pointers
-    EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, 0, n, (T) nullptr, lda, stA,
-                                                (U) nullptr, stP, dinfo, bc),
+    EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, 0, n, (Td) nullptr, lda,
+                                                stA, (Id) nullptr, stP, dInfo, bc),
                           rocblas_status_success);
-    EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, 0, (T) nullptr, lda, stA,
-                                                (U) nullptr, stP, dinfo, bc),
+    EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, 0, (Td) nullptr, lda,
+                                                stA, (Id) nullptr, stP, dInfo, bc),
                           rocblas_status_success);
     if(STRIDED)
         EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, dA, lda, stA,
-                                                    dIpiv, stP, (U) nullptr, 0),
+                                                    dIpiv, stP, (Id) nullptr, 0),
                               rocblas_status_success);
 
     // quick return with zero batch_count if applicable
     if(STRIDED)
         EXPECT_ROCBLAS_STATUS(
-            rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, dA, lda, stA, dIpiv, stP, dinfo, 0),
+            rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, dA, lda, stA, dIpiv, stP, dInfo, 0),
             rocblas_status_success);
 }
 
-template <bool BATCHED, bool STRIDED, bool GETRF, typename T>
+template <bool BATCHED, bool STRIDED, bool GETRF, typename T, typename I>
 void testing_getf2_getrf_bad_arg()
 {
     // safe arguments
     rocblas_local_handle handle;
-    rocblas_int m = 1;
-    rocblas_int n = 1;
-    rocblas_int lda = 1;
+    I m = 1;
+    I n = 1;
+    I lda = 1;
     rocblas_stride stA = 1;
     rocblas_stride stP = 1;
-    rocblas_int bc = 1;
+    I bc = 1;
 
     if(BATCHED)
     {
         // memory allocations
         device_batch_vector<T> dA(1, 1, 1);
-        device_strided_batch_vector<rocblas_int> dIpiv(1, 1, 1, 1);
-        device_strided_batch_vector<rocblas_int> dInfo(1, 1, 1, 1);
+        device_strided_batch_vector<I> dIpiv(1, 1, 1, 1);
+        device_strided_batch_vector<I> dInfo(1, 1, 1, 1);
         CHECK_HIP_ERROR(dA.memcheck());
         CHECK_HIP_ERROR(dIpiv.memcheck());
         CHECK_HIP_ERROR(dInfo.memcheck());
@@ -121,8 +121,8 @@ void testing_getf2_getrf_bad_arg()
     {
         // memory allocations
         device_strided_batch_vector<T> dA(1, 1, 1, 1);
-        device_strided_batch_vector<rocblas_int> dIpiv(1, 1, 1, 1);
-        device_strided_batch_vector<rocblas_int> dInfo(1, 1, 1, 1);
+        device_strided_batch_vector<I> dIpiv(1, 1, 1, 1);
+        device_strided_batch_vector<I> dInfo(1, 1, 1, 1);
         CHECK_HIP_ERROR(dA.memcheck());
         CHECK_HIP_ERROR(dIpiv.memcheck());
         CHECK_HIP_ERROR(dInfo.memcheck());
@@ -133,20 +133,19 @@ void testing_getf2_getrf_bad_arg()
     }
 }
 
-template <bool CPU, bool GPU, typename T, typename Td, typename Ud, typename Th, typename Uh>
+template <bool CPU, bool GPU, typename T, typename I, typename Td, typename Id, typename Th, typename Uh>
 void getf2_getrf_initData(const rocblas_handle handle,
-                          const rocblas_int m,
-                          const rocblas_int n,
+                          const I m,
+                          const I n,
                           Td& dA,
-                          const rocblas_int lda,
+                          const I lda,
                           const rocblas_stride stA,
-                          Ud& dIpiv,
+                          Id& dIpiv,
                           const rocblas_stride stP,
-                          Ud& dInfo,
-                          const rocblas_int bc,
+                          Id& dInfo,
+                          const I bc,
                           Th& hA,
                           Uh& hIpiv,
-                          Uh& hInfo,
                           const bool singular)
 {
     if(CPU)
@@ -154,12 +153,12 @@ void getf2_getrf_initData(const rocblas_handle handle,
         T tmp;
         rocblas_init<T>(hA, true);
 
-        for(rocblas_int b = 0; b < bc; ++b)
+        for(I b = 0; b < bc; ++b)
         {
             // scale A to avoid singularities
-            for(rocblas_int i = 0; i < m; i++)
+            for(I i = 0; i < m; i++)
             {
-                for(rocblas_int j = 0; j < n; j++)
+                for(I j = 0; j < n; j++)
                 {
                     if(i == j)
                         hA[b][i + j * lda] += 400;
@@ -170,9 +169,9 @@ void getf2_getrf_initData(const rocblas_handle handle,
 
             // shuffle rows to test pivoting
             // always the same permuation for debugging purposes
-            for(rocblas_int i = 0; i < m / 2; i++)
+            for(I i = 0; i < m / 2; i++)
             {
-                for(rocblas_int j = 0; j < n; j++)
+                for(I j = 0; j < n; j++)
                 {
                     tmp = hA[b][i + j * lda];
                     hA[b][i + j * lda] = hA[b][m - 1 - i + j * lda];
@@ -186,17 +185,17 @@ void getf2_getrf_initData(const rocblas_handle handle,
                 // (always the same elements for debugging purposes).
                 // The algorithm must detect the first zero pivot in those
                 // matrices in the batch that are singular
-                rocblas_int j = n / 4 + b;
+                I j = n / 4 + b;
                 j -= (j / n) * n;
-                for(rocblas_int i = 0; i < m; i++)
+                for(I i = 0; i < m; i++)
                     hA[b][i + j * lda] = 0;
                 j = n / 2 + b;
                 j -= (j / n) * n;
-                for(rocblas_int i = 0; i < m; i++)
+                for(I i = 0; i < m; i++)
                     hA[b][i + j * lda] = 0;
                 j = n - 1 + b;
                 j -= (j / n) * n;
-                for(rocblas_int i = 0; i < m; i++)
+                for(I i = 0; i < m; i++)
                     hA[b][i + j * lda] = 0;
             }
         }
@@ -209,29 +208,29 @@ void getf2_getrf_initData(const rocblas_handle handle,
     }
 }
 
-template <bool STRIDED, bool GETRF, typename T, typename Td, typename Ud, typename Th, typename Uh>
+template <bool STRIDED, bool GETRF, typename T, typename I, typename Td, typename Id, typename Th, typename Ih, typename Uh>
 void getf2_getrf_getError(const rocblas_handle handle,
-                          const rocblas_int m,
-                          const rocblas_int n,
+                          const I m,
+                          const I n,
                           Td& dA,
-                          const rocblas_int lda,
+                          const I lda,
                           const rocblas_stride stA,
-                          Ud& dIpiv,
+                          Id& dIpiv,
                           const rocblas_stride stP,
-                          Ud& dInfo,
-                          const rocblas_int bc,
+                          Id& dInfo,
+                          const I bc,
                           Th& hA,
                           Th& hARes,
                           Uh& hIpiv,
-                          Uh& hIpivRes,
+                          Ih& hIpivRes,
                           Uh& hInfo,
-                          Uh& hInfoRes,
+                          Ih& hInfoRes,
                           double* max_err,
                           const bool singular)
 {
     // input data initialization
     getf2_getrf_initData<true, true, T>(handle, m, n, dA, lda, stA, dIpiv, stP, dInfo, bc, hA,
-                                        hIpiv, hInfo, singular);
+                                        hIpiv, singular);
 
     // execute computations
     // GPU lapack
@@ -242,7 +241,7 @@ void getf2_getrf_getError(const rocblas_handle handle,
     CHECK_HIP_ERROR(hInfoRes.transfer_from(dInfo));
 
     // CPU lapack
-    for(rocblas_int b = 0; b < bc; ++b)
+    for(I b = 0; b < bc; ++b)
     {
         GETRF ? cpu_getrf(m, n, hA[b], lda, hIpiv[b], hInfo[b])
               : cpu_getf2(m, n, hA[b], lda, hIpiv[b], hInfo[b]);
@@ -255,14 +254,14 @@ void getf2_getrf_getError(const rocblas_handle handle,
     // using frobenius norm
     double err;
     *max_err = 0;
-    for(rocblas_int b = 0; b < bc; ++b)
+    for(I b = 0; b < bc; ++b)
     {
         err = norm_error('F', m, n, lda, hA[b], hARes[b]);
         *max_err = err > *max_err ? err : *max_err;
 
         // also check pivoting (count the number of incorrect pivots)
         err = 0;
-        for(rocblas_int i = 0; i < min(m, n); ++i)
+        for(I i = 0; i < min(m, n); ++i)
         {
             EXPECT_EQ(hIpiv[b][i], hIpivRes[b][i]) << "where b = " << b << ", i = " << i;
             if(hIpiv[b][i] != hIpivRes[b][i])
@@ -273,7 +272,7 @@ void getf2_getrf_getError(const rocblas_handle handle,
 
     // also check info for singularities
     err = 0;
-    for(rocblas_int b = 0; b < bc; ++b)
+    for(I b = 0; b < bc; ++b)
     {
         EXPECT_EQ(hInfo[b][0], hInfoRes[b][0]) << "where b = " << b;
         if(hInfo[b][0] != hInfoRes[b][0])
@@ -282,23 +281,23 @@ void getf2_getrf_getError(const rocblas_handle handle,
     *max_err += err;
 }
 
-template <bool STRIDED, bool GETRF, typename T, typename Td, typename Ud, typename Th, typename Uh>
+template <bool STRIDED, bool GETRF, typename T, typename I, typename Td, typename Id, typename Th, typename Uh>
 void getf2_getrf_getPerfData(const rocblas_handle handle,
-                             const rocblas_int m,
-                             const rocblas_int n,
+                             const I m,
+                             const I n,
                              Td& dA,
-                             const rocblas_int lda,
+                             const I lda,
                              const rocblas_stride stA,
-                             Ud& dIpiv,
+                             Id& dIpiv,
                              const rocblas_stride stP,
-                             Ud& dInfo,
-                             const rocblas_int bc,
+                             Id& dInfo,
+                             const I bc,
                              Th& hA,
                              Uh& hIpiv,
                              Uh& hInfo,
                              double* gpu_time_used,
                              double* cpu_time_used,
-                             const rocblas_int hot_calls,
+                             const int hot_calls,
                              const int profile,
                              const bool profile_kernels,
                              const bool perf,
@@ -307,11 +306,11 @@ void getf2_getrf_getPerfData(const rocblas_handle handle,
     if(!perf)
     {
         getf2_getrf_initData<true, false, T>(handle, m, n, dA, lda, stA, dIpiv, stP, dInfo, bc, hA,
-                                             hIpiv, hInfo, singular);
+                                             hIpiv, singular);
 
         // cpu-lapack performance (only if not in perf mode)
         *cpu_time_used = get_time_us_no_sync();
-        for(rocblas_int b = 0; b < bc; ++b)
+        for(I b = 0; b < bc; ++b)
         {
             GETRF ? cpu_getrf(m, n, hA[b], lda, hIpiv[b], hInfo[b])
                   : cpu_getf2(m, n, hA[b], lda, hIpiv[b], hInfo[b]);
@@ -320,13 +319,13 @@ void getf2_getrf_getPerfData(const rocblas_handle handle,
     }
 
     getf2_getrf_initData<true, false, T>(handle, m, n, dA, lda, stA, dIpiv, stP, dInfo, bc, hA,
-                                         hIpiv, hInfo, singular);
+                                         hIpiv, singular);
 
     // cold calls
     for(int iter = 0; iter < 2; iter++)
     {
         getf2_getrf_initData<false, true, T>(handle, m, n, dA, lda, stA, dIpiv, stP, dInfo, bc, hA,
-                                             hIpiv, hInfo, singular);
+                                             hIpiv, singular);
 
         CHECK_ROCBLAS_ERROR(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, dA.data(), lda, stA,
                                                   dIpiv.data(), stP, dInfo.data(), bc));
@@ -347,10 +346,10 @@ void getf2_getrf_getPerfData(const rocblas_handle handle,
         rocsolver_log_set_max_levels(profile);
     }
 
-    for(rocblas_int iter = 0; iter < hot_calls; iter++)
+    for(int iter = 0; iter < hot_calls; iter++)
     {
         getf2_getrf_initData<false, true, T>(handle, m, n, dA, lda, stA, dIpiv, stP, dInfo, bc, hA,
-                                             hIpiv, hInfo, singular);
+                                             hIpiv, singular);
 
         start = get_time_us_sync(stream);
         rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, dA.data(), lda, stA, dIpiv.data(), stP,
@@ -360,19 +359,19 @@ void getf2_getrf_getPerfData(const rocblas_handle handle,
     *gpu_time_used /= hot_calls;
 }
 
-template <bool BATCHED, bool STRIDED, bool GETRF, typename T>
+template <bool BATCHED, bool STRIDED, bool GETRF, typename T, typename I>
 void testing_getf2_getrf(Arguments& argus)
 {
     // get arguments
     rocblas_local_handle handle;
-    rocblas_int m = argus.get<rocblas_int>("m");
-    rocblas_int n = argus.get<rocblas_int>("n", m);
-    rocblas_int lda = argus.get<rocblas_int>("lda", m);
+    I m = argus.get<rocblas_int>("m");
+    I n = argus.get<rocblas_int>("n", m);
+    I lda = argus.get<rocblas_int>("lda", m);
     rocblas_stride stA = argus.get<rocblas_stride>("strideA", lda * n);
     rocblas_stride stP = argus.get<rocblas_stride>("strideP", min(m, n));
 
-    rocblas_int bc = argus.batch_count;
-    rocblas_int hot_calls = argus.iters;
+    I bc = argus.batch_count;
+    int hot_calls = argus.iters;
 
     rocblas_stride stARes = (argus.unit_check || argus.norm_check) ? stA : 0;
     rocblas_stride stPRes = (argus.unit_check || argus.norm_check) ? stP : 0;
@@ -393,14 +392,13 @@ void testing_getf2_getrf(Arguments& argus)
     if(invalid_size)
     {
         if(BATCHED)
-            EXPECT_ROCBLAS_STATUS(
-                rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, (T* const*)nullptr, lda, stA,
-                                      (rocblas_int*)nullptr, stP, (rocblas_int*)nullptr, bc),
-                rocblas_status_invalid_size);
+            EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n,
+                                                        (T* const*)nullptr, lda, stA, (I*)nullptr,
+                                                        stP, (I*)nullptr, bc),
+                                  rocblas_status_invalid_size);
         else
             EXPECT_ROCBLAS_STATUS(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, (T*)nullptr,
-                                                        lda, stA, (rocblas_int*)nullptr, stP,
-                                                        (rocblas_int*)nullptr, bc),
+                                                        lda, stA, (I*)nullptr, stP, (I*)nullptr, bc),
                                   rocblas_status_invalid_size);
 
         if(argus.timing)
@@ -415,12 +413,10 @@ void testing_getf2_getrf(Arguments& argus)
         CHECK_ROCBLAS_ERROR(rocblas_start_device_memory_size_query(handle));
         if(BATCHED)
             CHECK_ALLOC_QUERY(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, (T* const*)nullptr,
-                                                    lda, stA, (rocblas_int*)nullptr, stP,
-                                                    (rocblas_int*)nullptr, bc));
+                                                    lda, stA, (I*)nullptr, stP, (I*)nullptr, bc));
         else
             CHECK_ALLOC_QUERY(rocsolver_getf2_getrf(STRIDED, GETRF, handle, m, n, (T*)nullptr, lda,
-                                                    stA, (rocblas_int*)nullptr, stP,
-                                                    (rocblas_int*)nullptr, bc));
+                                                    stA, (I*)nullptr, stP, (I*)nullptr, bc));
 
         size_t size;
         CHECK_ROCBLAS_ERROR(rocblas_stop_device_memory_size_query(handle, &size));
@@ -439,12 +435,12 @@ void testing_getf2_getrf(Arguments& argus)
         host_batch_vector<T> hA(size_A, 1, bc);
         host_batch_vector<T> hARes(size_ARes, 1, bc);
         host_strided_batch_vector<rocblas_int> hIpiv(size_P, 1, stP, bc);
-        host_strided_batch_vector<rocblas_int> hIpivRes(size_PRes, 1, stPRes, bc);
+        host_strided_batch_vector<I> hIpivRes(size_PRes, 1, stPRes, bc);
         host_strided_batch_vector<rocblas_int> hInfo(1, 1, 1, bc);
-        host_strided_batch_vector<rocblas_int> hInfoRes(1, 1, 1, bc);
+        host_strided_batch_vector<I> hInfoRes(1, 1, 1, bc);
         device_batch_vector<T> dA(size_A, 1, bc);
-        device_strided_batch_vector<rocblas_int> dIpiv(size_P, 1, stP, bc);
-        device_strided_batch_vector<rocblas_int> dInfo(1, 1, 1, bc);
+        device_strided_batch_vector<I> dIpiv(size_P, 1, stP, bc);
+        device_strided_batch_vector<I> dInfo(1, 1, 1, bc);
         if(size_A)
             CHECK_HIP_ERROR(dA.memcheck());
         CHECK_HIP_ERROR(dInfo.memcheck());
@@ -483,12 +479,12 @@ void testing_getf2_getrf(Arguments& argus)
         host_strided_batch_vector<T> hA(size_A, 1, stA, bc);
         host_strided_batch_vector<T> hARes(size_ARes, 1, stARes, bc);
         host_strided_batch_vector<rocblas_int> hIpiv(size_P, 1, stP, bc);
-        host_strided_batch_vector<rocblas_int> hIpivRes(size_PRes, 1, stPRes, bc);
+        host_strided_batch_vector<I> hIpivRes(size_PRes, 1, stPRes, bc);
         host_strided_batch_vector<rocblas_int> hInfo(1, 1, 1, bc);
-        host_strided_batch_vector<rocblas_int> hInfoRes(1, 1, 1, bc);
+        host_strided_batch_vector<I> hInfoRes(1, 1, 1, bc);
         device_strided_batch_vector<T> dA(size_A, 1, stA, bc);
-        device_strided_batch_vector<rocblas_int> dIpiv(size_P, 1, stP, bc);
-        device_strided_batch_vector<rocblas_int> dInfo(1, 1, 1, bc);
+        device_strided_batch_vector<I> dIpiv(size_P, 1, stP, bc);
+        device_strided_batch_vector<I> dInfo(1, 1, 1, bc);
         if(size_A)
             CHECK_HIP_ERROR(dA.memcheck());
         CHECK_HIP_ERROR(dInfo.memcheck());
@@ -580,4 +576,5 @@ INSTANTIATE(EXTERN_TESTING_GETF2_GETRF,
             FOREACH_MATRIX_DATA_LAYOUT,
             FOREACH_BLOCKED_VARIANT,
             FOREACH_SCALAR_TYPE,
+            FOREACH_INT_TYPE,
             APPLY_STAMP)