+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Full documentation for rocSOLVER is available at the [rocSOLVER documentation](h
 ## (Unreleased) rocSOLVER
 ### Added
 - 64-bit APIs for existing functions:
+    - GETF2_64 (with batched and strided\_batched versions)
     - GETRS_64 (with batched and strided\_batched versions)
 
 ### Optimized